Force remove intermediate files
Intermediate files are created in R's `tempdir()` which will be attempted
to be deleted regardless of whether `clean` was TRUE or FALSE. When
these files were generated with read-only permissions, the delete
operation fails, and the markdown rendering exits with a non-zero exit
code.

Intermediate files are typically generated with read-only mode when HTML
dependencies are copied from the installed rmarkdown package if the
package installation had all files installed as read-only. A user can
install packages with read-only permissions when they want to be sure
that they are not accidentally modifying their R package versions, etc.